+++ src/test/java/com/iuliandogariu/linguist/webservice/NgramCountResponseTest.java
@@ -11,9 +11,9 @@ public final class NgramCountResponseTest {
 
     @Test
     public void shouldBuildFromCount() {
-        Map<Ngram, Integer> counts = new HashMap<>();
-        counts.put(new Ngram(3, "I love you"), 69);
-        Map.Entry<Ngram, Integer> count = counts.entrySet().iterator().next();
+        Map<Ngram, Long> counts = new HashMap<>();
+        counts.put(new Ngram(3, "I love you"), 69L);
+        Map.Entry<Ngram, Long> count = counts.entrySet().iterator().next();
         NgramCountResponse resp = NgramCountResponse.fromNgramCount(count);
         assertEquals("I love you", resp.getNgram());
         assertEquals(69, resp.getCount());
